added hook for climate personality
