+++ src/pages/PersonalValues.tsx
@@ -5,6 +5,7 @@ import { ReactComponent as Logo } from '../assets/cm-logo.svg';
 import ROUTES from '../components/Router/RouteConfig';
 
 import CMCard from '../components/CMCard';
+import { useClimatePersonality } from '../hooks/useClimatePersonality';
 
 const styles = makeStyles({
   root: {
@@ -18,23 +19,8 @@ const styles = makeStyles({
 });
 
 // TODO - Replace with values from the api
-const climatePersonality = [
-  {
-    valueDesc:
-      'What is important to you is the safety, harmony and stability of society, of relationships, and of self. Security values derive from basic individual and group needs. You value a sense of belonging, social order and the reciprocation of favours.',
-    valueName: 'security',
-  },
-  {
-    valueDesc:
-      "For you, respect, commitment and acceptance of the customs and ideas that one's culture or religion provides is highly important. It’s likely you practise a form of religious rites and beliefs. You are humble, devout and accepting of your portion in life.",
-    valueName: 'tradition',
-  },
-  {
-    valueDesc:
-      'You are excellent at restraint of actions, inclinations, and impulses likely to upset or harm others and violate social expectations or norms. Conformity values derive from the requirement that individuals inhibit inclinations that might disrupt and undermine smooth interaction and group functioning. You are obedient, self-disciplined, loyal, responsible and polite.',
-    valueName: 'conformity',
-  },
-];
+//const climatePersonality = useClimatePersonality();
+const climatePersonality = useClimatePersonality();
 
 const PersonalValues: React.FC = () => {
   const classes = styles();
